renamed Pages folder to Routes, set up route for table page
