+++ src/App.js
@@ -2,16 +2,12 @@ import React, { Component } from 'react'
 import { BrowserRouter, Route, Redirect } from 'react-router-dom'
 import MediaQuery from 'react-responsive'
 
-// import Nav from './Pages/Interface/Nav'
-// import MobileNav from './Pages/Interface/MobileNav'
-import Home from './Pages/Home'
-import Archive from './Pages/Archive'
-import User from './Pages/User'
-import Draft from './Draft/Draft'
-import DesktopRegister from './Pages/DesktopRegister'
-import MobileRegister from './Pages/MobileRegister'
-
-import './App.css'
+// import Nav from './Routes/Interface/Nav'
+// import MobileNav from './Routes/Interface/MobileNav'
+// import Draft from './Draft/Draft'
+import RegisteredPlayers from './Routes/RegisteredPlayers'
+import DesktopRegister from './Routes/DesktopRegister'
+import MobileRegister from './Routes/MobileRegister'
 
 import { MuiThemeProvider, createMuiTheme } from '@material-ui/core/styles'
 import CssBaseline from '@material-ui/core/CssBaseline'
@@ -55,21 +51,19 @@ class App extends Component {
         <CssBaseline />
         <BrowserRouter>
           <div>
+            <Route exact path='/' render={() => <Redirect to="/register" /> } />
             {/* WEB */}
-            <Route exact path="/" render={() => <Redirect to="/register" /> } />
             <MediaQuery minDeviceWidth={1224}>
               {/* <Nav /> */}
-              <Route path="/home" component={Home} />
-              <Route path="/archive" component={Archive} />
-              <Route path="/user" component={User} />
-              <Route path="/draft" component={Draft} />
-              <Route path="/register" component={DesktopRegister} />
+              {/* <Route path='/draft' component={Draft} /> */}
+              <Route path='/players' component={RegisteredPlayers} />
+              <Route path='/register' component={DesktopRegister} />
             </MediaQuery>
 
             {/* MOBILE */}
             <MediaQuery maxDeviceWidth={1224}>
               {/* <MobileNav /> */}
-              <Route path="/register" component={MobileRegister} />
+              <Route path='/register' component={MobileRegister} />
             </MediaQuery>
           </div>
         </BrowserRouter>
