--- src/Pages/Archive.js
@@ -1,7 +1,0 @@
-import React from 'react'
-
-const Archive = (props) => {
-  return <div>Archive</div>
-}
-
-export default Archive
