--- src/Pages/Home.js
@@ -1,15 +1,0 @@
-import React, { Component } from 'react'
-
-import './MobileRegister.css'
-
-class Home extends Component {
-  render () {
-    return (
-      <div className='main'>
-
-      </div>
-    )
-  }
-}
-
-export default Home
