--- src/Pages/User.js
@@ -1,7 +1,0 @@
-import React from 'react'
-
-const User = (props) => {
-  return <div>User</div>
-}
-
-export default User
