+++ src/Routes/RegisteredPlayers.js
@@ -1,0 +1,15 @@
+import React, { Component } from 'react'
+
+import './PlayerList.css'
+
+class RegisteredPlayers extends Component {
+  render () {
+    return (
+      <div className='playerListDisplay'>
+
+      </div>
+    )
+  }
+}
+
+export default RegisteredPlayers
